+++ _example/custom_driver_name/main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"database/sql"
 
-	_ "github.com/mattn/go-sqlite3"
+	_ "github.com/meow-io/go-sqlcipher"
 )
 
 func main() {